# feat: optimize mints and swaps queries