# Summary

Optimize the `swaps` and `mints` queries by querying the two entities directly at the same time.

|Before |After |
--- | --- |
|2:27m|1:40m|
|2:10m|1:57m|
|2:15m|1:46m|

With these changes, the chart should load faster by at least ~20s.

# How To Test
1.  Open the page `dashboard`
- [ ] The `Wallets` history chart should load a bit faster